chore(jest): port CodeReview and FindProgrammingExpressionDialogTest
FindProgrammingExpressionDialogTest uses a fetch call which returns a rejected
promise as the API endpoint doesn't exist. In Karma this reject is silently
ignored but jest will error on them.

CodeReview uses references that results in a fetch call, again it is not
needed within the test because in Karma mode the fetch rejects and is
ignored.

Finally, the test results in a promise which needs to be returned in
jest.