chore(ejs): enable ejs javascript strict mode to remove with syntax
Note: This change is required for the jest migration as jest enables
strict mode during babel transpilation.

Currently, Code.org uses EJS to templatize files. EJS, by default,
utilizes the `with` keyword which is deprecated and per MDN, "this
feature may cease to work at any time.". To remove this keyword, the EJS
strict mode option is enabled (`with` is prohibited in strict).

To replace the functionality previously provided by `with` (add
variables within a given block scope), the variables are destructured
from the `locals` variable (which is provided by EJS). In strict mode,
variables are not scoped globally by default and therefore assigning the
variables within the EJS strict mode provides the same functionality as
`with` but with extra syntax needed to use an EJS template.

Previously, the code would be able to function:

```
<% hello %>
```

Now with strict mode enabled, the same code would need to become:

```
<% locals.hello %>
```

To avoid mistakes, a variable reassignment block was created to existing
EJS files. This will enable the variables to continue to be used as if
the with block was available.